refactor(k6): rename run.ts to main.ts
Align file name with exported function.

Session-ID: unavailable
